rename file: avoid namespace issues
